requirements.txt
flask
confluent-kafka
python-json-logger
python-dotenv
requests
gunicorn
